Encapsulate and simplify Semantic Analysis
