+++ web/src/types/imports-components.d.ts
@@ -2,7 +2,7 @@
 // We suggest you to commit this file into source control
 // Read more: https://github.com/vuejs/core/pull/3399
 export {}
-        
+
 declare module 'vue' {
   export interface GlobalComponents {
     Breadcrumbs: typeof import('./../components/Breadcrumbs.vue')['default']
